#62 - moved Logout button to top right corner of main app layout

+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.css';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import React from 'react';
-import {Switch, Route, RouteComponentProps} from 'react-router-dom';
+import {Switch, Route, RouteComponentProps, Link} from 'react-router-dom';
 import {Tab, Box, Paper} from '@mui/material';
 import {TabList, TabPanel, TabContext} from '@mui/lab';
 //import { Patient} from './data-services/fhir-types/fhir-r4';
@@ -37,6 +37,7 @@ import {
     getSelectedEndpoints,
     deleteSelectedEndpoints,
     getLauncherData,
+    launcherDataExists,
     deleteAllDataFromLocalForage,
     saveSessionId,
     sessionIdExistsInLocalForage,
@@ -853,14 +854,15 @@ class App extends React.Component<AppProps, AppState> {
         this.setState({goalSummaries: newGoalSummaries})
     }
 
-    // setLogout called when un-sharing data to the SDS
-    setLogout = () => {
-        this.setState({isLogout: true});
-        sessionStorage.clear();
-        deleteAllDataFromLocalForage();
-        this.props.history.push('/logout')
-        // this.setState({ isLoggedOut: true });
-    }
+    handleLogout = async () => {
+        if (await launcherDataExists()) {
+            console.log('Logging out...')
+            this.setState({isLogout: true});
+            sessionStorage.clear();
+            await deleteAllDataFromLocalForage();
+            this.props.history.push('/logout')
+        }
+    };
 
     // callback function to update conditions from ConditionEditForm
     setConditionSummaries = (newConditionSummaries: ConditionSummary[][]) => {
@@ -1008,19 +1010,6 @@ class App extends React.Component<AppProps, AppState> {
         this.setState({userErrorMessage: undefined})
     }
 
-    // storer: commenting out handleLogout and slating for deletion, as it does literally nothing
-    // private handleLogout = async () => {
-    //     if (!this.state.isLogout) {
-    //         // Clear session storage or perform other logout logic here
-    //         // console.log('Logging out and clearing session');
-    //         // this.setState({ isLogout: true });
-    //         // this.setState({ isLogout: true })
-    //         // sessionStorage.clear()
-    //         // await deleteAllDataFromLocalForage()
-    //         // this.props.history.push('/logout')
-    //     }
-    // }
-
     updateLogMainTab = async (event: any, value: any) => {
         this.setState({mainTabIndex: value});
 
@@ -1112,8 +1101,18 @@ class App extends React.Component<AppProps, AppState> {
                     {/* <img className="mypain-header-logo" src={`${process.env.PUBLIC_URL}/assets/images/mpc-logo.png`} alt="MyPreventiveCare"/> */}
                     <img className="mypain-header-logo"
                          src={`${process.env.PUBLIC_URL}/assets/images/ecareplan-logo.png`} alt="My Care Planner"/>
+                    <span style={{margin: 'auto' }}>&nbsp;</span>
                     {/* {patient === undefined ? '' : <p>&npsp;&npsp;{patient[0]?.fullName}</p>} */}
-                    <p className='version'>{process.env.REACT_APP_VERSION}</p>
+
+                    <div className='app-header-right'>
+                        { ! this.state.isLogout ?
+                            <span><Link to="/logout" onClick={this.handleLogout} className='logoutLink'>
+                                Logout
+                            </Link></span> : ''
+                        }
+                        <span className='version'>{process.env.REACT_APP_VERSION}</span>
+                    </div>
+
                 </header>
 
                 <Switch>
@@ -1146,11 +1145,11 @@ class App extends React.Component<AppProps, AppState> {
                                />
                            )}
                     />
-                    <Route path="/unshare-data">
-                        <SessionProtected isLoggedIn={!this.state.isLogout}>
-                            <UnShareData fhirDataCollection={this.state.fhirDataCollection} setLogout={this.setLogout}/>
-                        </SessionProtected>
-                    </Route>
+                    {/*<Route path="/unshare-data">*/}
+                    {/*    <SessionProtected isLoggedIn={!this.state.isLogout}>*/}
+                    {/*        <UnShareData fhirDataCollection={this.state.fhirDataCollection} setLogout={this.setLogout}/>*/}
+                    {/*    </SessionProtected>*/}
+                    {/*</Route>*/}
                     <Route path="/shared-data-summary">
                         <SessionProtected isLoggedIn={!this.state.isLogout}>
                             <SharedDataSummary/>
@@ -1299,7 +1298,7 @@ class App extends React.Component<AppProps, AppState> {
                                             <h4 className="title">Assessment Results</h4>
                                             <p>Coming soon...</p>
                                         </TabPanel> */}
-                                            <TabPanel value="11">
+                                            <TabPanel value="11" sx={{padding: '0px 15px'}}>
                                                 <ImmunizationList sharingData={this.state.sharingData}
                                                                   fhirDataCollection={this.state.fhirDataCollection}
                                                                   progressTitle={this.state.progressTitle}
